working with long timeout

+++ arduino/sketchbook/libraries/MQTTClient/Countdown.h
@@ -25,7 +25,7 @@ public:
 		interval_end_ms = 0L;
     }
     
-    Countdown(int ms)
+    Countdown(unsigned long int ms)
     {
         countdown_ms(ms);   
     }
@@ -35,23 +35,23 @@ public:
         return (interval_end_ms > 0L) && (millis() >= interval_end_ms);
     }
     
-    void countdown_ms(unsigned long ms)  
+    void countdown_ms(unsigned long int ms)  
     {
         interval_end_ms = millis() + ms;
     }
     
     void countdown(int seconds)
     {
-        countdown_ms((unsigned long)seconds * 1000L);
+        countdown_ms((unsigned long int)seconds * 1000L);
     }
     
-    int left_ms()
+    unsigned long int left_ms()
     {
         return interval_end_ms - millis();
     }
     
 private:
-    unsigned long interval_end_ms; 
+    unsigned long int interval_end_ms; 
 };
 
 #endif
